autosteer_server/pyproject.toml
[project]
name = "autosteer"
version = "0.1.0"
description = "Autosteer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.34.2",
    "modal>=0.64.139",
    "python-dotenv>=1.0.1",
    "torch>=2.4.1",
    "transformers>=4.44.2",
]
